feat : "도서 상세 페이지 mapping"
